requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
docx2python==2.7.3
fonttools==4.41.1
frozenlist==1.4.0
google-search-results==2.4.2
idna==3.4
joblib==1.3.1
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.7.2
multidict==6.0.4
nltk==3.8.1
numpy==1.25.1
openai==0.27.8
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.15.0
pony==0.7.16
pyparsing==3.0.9
pypdf==3.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.4
uule-grabber==0.1.9
yarl==1.9.2
